+++ src/components/common/CollectionBanner/index.tsx
