+++ src/components/common/CollectionBanner/styles.json
